Rename CurrentDutyStationForm to CurrentDutyLocationForm
